批量清理warning(100+->67): Icons.AutoMirrored迁移(11处)+逻辑冗余删除(35处safe-call/elvis/死分支)

+++ app/src/main/java/top/hsyscn/opedrgent/stt/StepAudioAsrEngine.kt
@@ -143,7 +143,7 @@ class StepAudioAsrEngine(
             val processingTimeMs = System.currentTimeMillis() - startTimeMs
             result.copy(
                 processingTimeMs = processingTimeMs,
-                durationMs = file.length().toLong(),
+                durationMs = file.length(),
                 modelUsed = "stepaudio-2.5-asr",
             )
         } catch (e: Exception) {
